LPS-197075 Inline AttachmentUploadFileEntryHandler in ObjectDefinitionDeployerImpl
