+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/entries/portlet/action/UploadAttachmentMVCActionCommand.java
@@ -5,8 +5,8 @@
 
 package com.liferay.object.web.internal.object.entries.portlet.action;
 
-import com.liferay.object.web.internal.object.entries.upload.AttachmentUploadFileEntryHandler;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
+import com.liferay.upload.UploadFileEntryHandler;
 import com.liferay.upload.UploadHandler;
 import com.liferay.upload.UploadResponseHandler;
 
@@ -19,11 +19,11 @@ import javax.portlet.ActionResponse;
 public class UploadAttachmentMVCActionCommand extends BaseMVCActionCommand {
 
 	public UploadAttachmentMVCActionCommand(
-		AttachmentUploadFileEntryHandler attachmentUploadFileEntryHandler,
+		UploadFileEntryHandler uploadFileEntryHandler,
 		UploadResponseHandler uploadResponseHandler,
 		UploadHandler uploadHandler) {
 
-		_attachmentUploadFileEntryHandler = attachmentUploadFileEntryHandler;
+		_uploadFileEntryHandler = uploadFileEntryHandler;
 		_uploadResponseHandler = uploadResponseHandler;
 		_uploadHandler = uploadHandler;
 	}
@@ -34,14 +34,13 @@ public class UploadAttachmentMVCActionCommand extends BaseMVCActionCommand {
 		throws Exception {
 
 		_uploadHandler.upload(
-			_attachmentUploadFileEntryHandler, _uploadResponseHandler,
-			actionRequest, actionResponse);
+			_uploadFileEntryHandler, _uploadResponseHandler, actionRequest,
+			actionResponse);
 
 		hideDefaultSuccessMessage(actionRequest);
 	}
 
-	private final AttachmentUploadFileEntryHandler
-		_attachmentUploadFileEntryHandler;
+	private final UploadFileEntryHandler _uploadFileEntryHandler;
 	private final UploadHandler _uploadHandler;
 	private final UploadResponseHandler _uploadResponseHandler;
